requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
cupy-cuda11x>=11.0.0
cuquantum-python>=22.11.0
yfinance>=0.2.18
statsmodels>=0.13.5
matplotlib>=3.5.0
scikit-learn>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0